🔨🤖 move the marimekko files to their own folder
Marimekko stopped stacking anything when it dropped to a single y indicator, so
the files no longer belong in stackedCharts. It is not a bar chart either, so
give it src/marimekko/ rather than folding it into barCharts. A pure move plus
its import sites and the oxlint return-type override, which enumerates
directories. MarimekkoBars still reads STACKED_BAR_STYLE across the boundary
because the opacity ramp is genuinely shared with the stacked charts.
