+++ packages/@ourworldindata/grapher/src/marimekko/MarimekkoChartThumbnail.tsx
