+++ packages/@ourworldindata/grapher/src/marimekko/MarimekkoInternalLabels.tsx
